Move QNameMapping to its own source file and implement missing GetHashCode method.
