+++ Libraries/dotNetRDF/Core/NamespaceMapper.cs
@@ -625,75 +625,4 @@ namespace VDS.RDF
             }
         }
     }
-
-    /// <summary>
-    /// Represents a mapping from a URI to a QName.
-    /// </summary>
-    public class QNameMapping 
-    {
-        private string _u;
-        private string _qname;
-
-        /// <summary>
-        /// Creates a new QName Mapping.
-        /// </summary>
-        /// <param name="u">URI.</param>
-        public QNameMapping(string u)
-        {
-            _u = u;
-        }
-
-        /// <summary>
-        /// URI this is a mapping for.
-        /// </summary>
-        public string Uri
-        {
-            get
-            {
-                return _u;
-            }
-        }
-
-        /// <summary>
-        /// QName this URI maps to.
-        /// </summary>
-        public string QName
-        {
-            get
-            {
-                return _qname;
-            }
-            set
-            {
-                _qname = value;
-            }
-        }
-
-        /// <summary>
-        /// Gets the String representation of the URI.
-        /// </summary>
-        /// <returns></returns>
-        public override string ToString()
-        {
-            return _u.ToString();
-        }
-
-        /// <summary>
-        /// Checks whether this is equal to another Object.
-        /// </summary>
-        /// <param name="obj">Object to test against.</param>
-        /// <returns></returns>
-        public override bool Equals(object obj)
-        {
-            if (obj == null) return false;
-            if (obj is QNameMapping)
-            {
-                return ToString().Equals(obj.ToString(), StringComparison.Ordinal);
-            }
-            else
-            {
-                return false;
-            }
-        }
-    }
 }
